# swagger.yaml
openapi: 3.0.3
info:
  title: To-do app API
  description: API for managing your tasks
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /create-task:
    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                due_time:
                  type: timestamp
                user_id:
                  type: integer
      responses:
        '201':
          description: Task created succesfully
        '400':
          description: Bad request
        '401':
          description: Not authorized
...